refactor: Distance 패키지 변경 및 예외 처리 메시지 작성
