+++ src/main/java/nextstep/subway/DataLoaderConfig.java
@@ -1,7 +1,7 @@
 package nextstep.subway;
 
 import com.google.common.collect.Lists;
-import nextstep.subway.line.application.Distance;
+import nextstep.subway.line.domain.Distance;
 import nextstep.subway.line.domain.Line;
 import nextstep.subway.line.domain.LineRepository;
 import nextstep.subway.member.domain.Member;
