+++ src/main/java/nextstep/subway/line/application/LineService.java
@@ -1,5 +1,6 @@
 package nextstep.subway.line.application;
 
+import nextstep.subway.line.domain.Distance;
 import nextstep.subway.line.domain.Line;
 import nextstep.subway.line.domain.LineRepository;
 import nextstep.subway.line.domain.Section;
@@ -47,8 +48,7 @@ public class LineService {
   }
 
   public Line findLineById(Long id) {
-    return lineRepository.findById(id)
-            .orElseThrow(RuntimeException::new);
+    return findLine(id);
   }
 
 
@@ -60,7 +60,7 @@ public class LineService {
 
   @Transactional
   public void updateLine(Long id, LineRequest lineUpdateRequest) {
-    Line persistLine = lineRepository.findById(id).orElseThrow(RuntimeException::new);
+    Line persistLine = findLine(id);
     persistLine.update(new Line(lineUpdateRequest.getName(), lineUpdateRequest.getColor()));
   }
 
@@ -82,5 +82,9 @@ public class LineService {
     Line line = findLineById(lineId);
     Station station = stationService.findStationById(stationId);
     line.removeStation(station);
+  }
+
+  private Line findLine(Long id) {
+    return lineRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("존재하지 않은 지하철 역입니다."));
   }
 }
