+++ src/main/java/nextstep/subway/line/domain/Distance.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line.application;
+package nextstep.subway.line.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
