+++ src/test/java/nextstep/subway/line/acceptance/SectionsTest.java
@@ -1,6 +1,6 @@
 package nextstep.subway.line.acceptance;
 
-import nextstep.subway.line.application.Distance;
+import nextstep.subway.line.domain.Distance;
 import nextstep.subway.line.domain.Line;
 import nextstep.subway.line.domain.Section;
 import nextstep.subway.station.domain.Station;
