+++ src/test/java/study/unit/UnitTest.java
@@ -1,6 +1,6 @@
 package study.unit;
 
-import nextstep.subway.line.application.Distance;
+import nextstep.subway.line.domain.Distance;
 import nextstep.subway.line.domain.Line;
 import nextstep.subway.station.domain.Station;
 import org.junit.jupiter.api.DisplayName;
